requirements.txt
streamlit_lottie
streamlit-option-menu
streamlit
google-generativeai
docx
json